+++ ihmc-perception/src/mapsense-wrapper/cpp/geometry/keyframe.h
@@ -7,18 +7,18 @@ struct Keyframe
 {
    public:
 
-      Keyframe(uint32_t kid, Eigen::Matrix4f pose, cv::Mat desc, KeyPointVec kp, const std::vector<int>& ids, cv::Mat left)
+      Keyframe(uint32_t kid, Eigen::Matrix4d pose, cv::Mat desc, KeyPointVec kp, const std::vector<int>& ids, cv::Mat left)
          : id(kid), pose(pose), descLeft(desc), keypointsLeft(kp), keypointIDs(ids), leftImage(left)  {};
 
-      Keyframe(uint32_t kid, Eigen::Matrix4f pose, cv::Mat descLeft, cv::Mat descRight, KeyPointVec kpLeft, KeyPointVec kpRight, cv::Mat left, cv::Mat right)
+      Keyframe(uint32_t kid, Eigen::Matrix4d pose, cv::Mat descLeft, cv::Mat descRight, KeyPointVec kpLeft, KeyPointVec kpRight, cv::Mat left, cv::Mat right)
             : id(kid), pose(pose), descLeft(descLeft), descRight(descRight), keypointsLeft(kpLeft), keypointsRight(kpRight), leftImage(left), rightImage(right) {};
       
-      Keyframe(uint32_t kid, Eigen::Matrix4f pose, cv::Mat descLeft, cv::Mat descRight, KeyPointVec kpLeft, KeyPointVec kpRight)
+      Keyframe(uint32_t kid, Eigen::Matrix4d pose, cv::Mat descLeft, cv::Mat descRight, KeyPointVec kpLeft, KeyPointVec kpRight)
             : id(kid), pose(pose), descLeft(descLeft), descRight(descRight), keypointsLeft(kpLeft), keypointsRight(kpRight) {};
       
       uint32_t id = 0;
 
-      Eigen::Matrix4f pose;
+      Eigen::Matrix4d pose;
 
       cv::Mat descLeft;
       cv::Mat descRight;
