+++ ihmc-perception/src/slam-wrapper/cpp/external/FactorGraphExternal.h
@@ -12,7 +12,7 @@ class FactorGraphExternal
         void addOdometryFactor(float* odometry, int poseId);
 
         // Expects 4x4 homogenous transform matrix to insert Pose3 factor
-        void addOdometryFactorExtended(float *odometry, int poseId);
+        void addOdometryFactorExtended(double *odometry, int poseId);
 
         // Expects packed Vector4
         void addOrientedPlaneFactor(float* lmMean, int lmId, int poseIndex);
